Implement panning and zooming for Arc Diagram (#65)
* Add initial drawing code for arcdiagram

* Separate arc diagram component into ArcDiagram.vue

* test commit

* Make the arc diagram work with a graph object

* test commit

* Implemented linear ArcDiagram
Added interactivity (highlighting for nodes)

* fixed a bug where arcs would be drawn wrong if sourceNode had higher index than targetNode

* Add tooltip feature

* Stole tooltip code from Wout and added it to linear arc diagram(thank you wout)

* graph interactivity is now on-hover (thank you Wout) instead of on-click (fixes some nasty bugs)

* Add initial drawing code for arcdiagram

* Separate arc diagram component into ArcDiagram.vue

* Make the arc diagram work with a graph object

* Add tooltip feature

* Store "index" and "text" in a separate map

Before, the index and text element assigned to a node was stored directly in the node in the Graph object. This prevented me from calling draw() a second time without causing problems; this should now be partially resolved

* Draw circular ArcDiagram with circle as node

* Combine mouseover and mouseout implementations of linear and circular impl

Instead of creating a new mouseover/mouseout event listener every time the VIS is drawn, it now does this at the start of load. It also stores the Graphics for each circle in the map instead of the graph's attributes.

* fix: vis is redrawn on window resize
NOTE: there is a bug when trying to englarge the window more than it was originally (idk how to fix that :( )

* Formatting and small issues fixed

* Move ArcDiagram to correct folder

* FIX: fixed bug with resizing

* FIX: edges that would overlap visually are now associated to the same graphics object
tweaked a bit the color scheme
add all edges(graphics obj) to a map with key of type string (i.e. 2-

* experimental for zooming and panning
some stuff with colors

* fixed dependecies

* FIX: highlighted edges are easier to visualise
zooming and panning
circle diagram radius

* changed something that might help with performance

* cleaned up the code by deleting useless comments

* Fix sunburst text drawing in corner

* Remove unused nodeMap

* Remove unnecessary code, change function name and add comments

Co-authored-by: Vlad Plaiasu <v.plaiasu@student.tue.nl>
Co-authored-by: Stijn <6910589+Stijn-H@users.noreply.github.com>
Co-authored-by: Radu Pogonariu <r.pogonariu@student.tue.nl>